Add pry and awesome_print as runtime dependencies
When these gems were added to the gemfile, they were mistakenly not
added as runtime dependencies. This meant the gem would not correctly
require them when installed through bundler. This was recorded in a
Github issue[1].

This commit adds `pry` and `awesome_print` as runtime dependencies so
they will be correctly required by bundler and the gem will run.

[1] - https://github.com/rusic/rusic-gem/issues/28